requirements.txt
black==23.1.0
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
python-dotenv==0.21.1
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
